+++ experimental-examples/tina-cloud-starter-iframe/components/blocks/content.tsx
@@ -1,19 +1,19 @@
 import React from "react";
-import { Container } from "../container";
-import { Section } from "../section";
+import { Container } from "../util/container";
+import { Section } from "../util/section";
 import { TinaMarkdown } from "tinacms/dist/rich-text";
 import type { TinaTemplate } from "tinacms";
-import { tinaField } from "tinacms/dist/react";
 
 export const Content = ({ data, parentField = "" }) => {
   return (
     <Section color={data.color}>
       <Container
-        className={`max-w-4xl prose prose-lg ${
+        className={`prose prose-lg ${
           data.color === "primary" ? `prose-primary` : `dark:prose-dark`
         }`}
-        data-tinafield={tinaField(data, "body")}
+        data-tinafield={`${parentField}.body`}
         size="large"
+        width="medium"
       >
         <TinaMarkdown content={data.body} />
       </Container>