+++ packages/@tinacms/graphql/src/resolver/index.ts
@@ -1,14 +1,5 @@
 /**
-Copyright 2021 Forestry.io Holdings, Inc.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+
 */
 
 import path from 'path'
@@ -32,13 +23,13 @@ import type {
 import type { GraphQLConfig } from '../types'
 
 import { TinaGraphQLError, TinaParseDocumentError } from './error'
-import { FilterCondition, makeFilterChain } from '@tinacms/datalayer'
 import { collectConditionsForField, resolveReferences } from './filter-utils'
 import {
   resolveMediaRelativeToCloud,
   resolveMediaCloudToRelative,
 } from './media-utils'
 import { GraphQLError } from 'graphql'
+import { FilterCondition, makeFilterChain } from '../database/datalayer'
 
 interface ResolverConfig {
   config?: GraphQLConfig