+++ packages/@tinacms/graphql/src/spec/movies-with-datalayer/requests.test.ts
@@ -1,20 +1,11 @@
 /**
-Copyright 2021 Forestry.io Holdings, Inc.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+
 */
 
 import path from 'path'
 import { setupFixture, setupFixture2, print, Fixture } from '../setup'
-import { LevelStore } from '@tinacms/datalayer'
 import { tinaSchema } from './.tina/schema'
+import { MemoryLevel } from 'memory-level'
 const rootPath = path.join(__dirname, '/')
 
 const fixtures: Fixture[] = [
@@ -77,16 +68,18 @@ afterEach(() => {
 })
 
 describe('A schema with indexing', () => {
-  let store
+  let level
   beforeEach(() => {
-    store = new LevelStore(rootPath, true)
+    level = new MemoryLevel<string, Record<string, any>>({
+      valueEncoding: 'json',
+    })
   })
   fixtures.forEach((fixture) => {
     it(print(fixture), async () => {
       const { responses, expectedResponsePaths } = await setupFixture(
         rootPath,
         tinaSchema,
-        store,
+        level,
         fixture,
         'movies-with-datalayer'
       )
@@ -110,7 +103,7 @@ describe('A schema with indexing', () => {
       const { responses, expectedResponsePaths } = await setupFixture2(
         rootPath,
         tinaSchema,
-        store,
+        level,
         fixture,
         'movies-with-datalayer',
         '_mutation',