+++ packages/@tinacms/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tinacms-cli
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [61f8c0e50]
+  - @tinacms/app@1.0.6
+
+## 1.0.8
+
+### Patch Changes
+
+- 55f7953eb: Small cli styling change on public assets in init
+  - @tinacms/app@1.0.5
+
 ## 1.0.7
 
 ### Patch Changes