+++ packages/tinacms/CHANGELOG.md
@@ -1,5 +1,12 @@
 # tinacms
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [09e716538]
+  - @tinacms/toolkit@1.3.1
+
 ## 1.1.4
 
 ### Patch Changes